pyproject.toml
[tool.poetry]
name = "ossprey"
version = "1.0.13"
description = "Ossprey tooling to scan your software package, create an SBOM and then submit it to our service for malware scanning"
authors = ["dreadn0ught"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
cyclonedx-bom = "^5.1.1"
cyclonedx-python-lib = {extras = ["json-validation"], version = "^8.4.0"}
requests = "^2.32.3"
pygithub = "^2.4.0"
packageurl-python = "^0.16.0"
ossbom = "^1.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.14.1"
twine = "^6.1.0"

[tool.poetry.scripts]
ossprey = "ossprey.scan:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
test = "poetry run pytest"
deploy = "poetry build && twine upload dist/*"
deploy-test = "poetry build && twine upload --repository-url https://test.pypi.org/legacy/ dist/*"
